Fixed more bugs with verification alg. Missing only caching for binary and ver cases
